+++ Assets/Novel/Tests/EditMode/Resources/Scenarios/test_unknown_command.rb
@@ -1,0 +1,2 @@
+mystery_effect 'sparkle'
+bg 'bg_after_unknown'